+++ playground/inc.js
@@ -1,11 +1,13 @@
-import React from 'react';
-import { styler } from '../packages/popmotion/lib';
+import React from "react";
+import { styler } from "../packages/popmotion/lib";
 
-const style = { width: '100px', height: '100px', background: 'red' };
-export const Box = ({ innerRef, ...props }) => <div ref={innerRef} {...props} style={{ ...props.style, ...style }} />;
+const style = { width: "100px", height: "100px", background: "red" };
+export const Box = ({ innerRef, ...props }) => (
+  <div ref={innerRef} {...props} style={{ ...style, ...props.style }} />
+);
 
 export class BaseAnimation extends React.Component {
-  setRef = (dom) => {
+  setRef = dom => {
     if (!dom) return;
     this.animation = this.getAnimation(styler(dom));
   };